python--ldap-list-users-data/requirements.txt
python-ldap
pyyaml
tabulate
